requirements.txt
mysql-connector-python~=8.4.0

streamlit-option-menu

sqlalchemy

streamlit~=1.34.0
pandas~=2.2.2